+++ game/transfers.py
@@ -228,37 +228,41 @@ class AirliftPlanner:
         distance_cache = ObjectiveDistanceCache.get_closest_airfields(
             self.transfer.position
         )
+        air_wing = self.game.air_wing_for(self.for_player)
         for cp in distance_cache.closest_airfields:
             if cp.captured != self.for_player:
                 continue
 
             inventory = self.game.aircraft_inventory.for_control_point(cp)
             for unit_type, available in inventory.all_aircraft:
-                squadrons = [
-                    s
-                    for s in self.game.air_wing_for(self.for_player).squadrons_for(
-                        unit_type
-                    )
-                    if FlightType.TRANSPORT in s.auto_assignable_mission_types
-                ]
-                if not squadrons:
-                    continue
-                squadron = squadrons[0]
-                if self.compatible_with_mission(unit_type, cp):
-                    while available and self.transfer.transport is None:
-                        flight_size = self.create_airlift_flight(squadron, inventory)
-                        available -= flight_size
+                squadrons = air_wing.auto_assignable_for_task_with_type(
+                    unit_type, FlightType.TRANSPORT
+                )
+                for squadron in squadrons:
+                    if self.compatible_with_mission(unit_type, cp):
+                        while (
+                            available
+                            and squadron.has_available_pilots
+                            and self.transfer.transport is None
+                        ):
+                            flight_size = self.create_airlift_flight(
+                                squadron, inventory
+                            )
+                            available -= flight_size
         if self.package.flights:
             self.game.ato_for(self.for_player).add_package(self.package)
 
     def create_airlift_flight(
         self, squadron: Squadron, inventory: ControlPointAircraftInventory
     ) -> int:
-        available = inventory.available(squadron.aircraft)
+        available_aircraft = inventory.available(squadron.aircraft)
         capacity_each = 1 if squadron.aircraft.dcs_unit_type.helicopter else 2
         required = math.ceil(self.transfer.size / capacity_each)
         flight_size = min(
-            required, available, squadron.aircraft.dcs_unit_type.group_size_max
+            required,
+            available_aircraft,
+            squadron.aircraft.dcs_unit_type.group_size_max,
+            squadron.number_of_available_pilots,
         )
         capacity = flight_size * capacity_each