+++ qt_ui/windows/AirWingDialog.py
@@ -48,10 +48,10 @@ class SquadronDelegate(TwoColumnRowDelegate):
             return self.squadron(index).nickname or ""
         elif (row, column) == (1, 1):
             squadron = self.squadron(index)
-            alive = squadron.number_of_living_pilots
             active = len(squadron.active_pilots)
             available = len(squadron.available_pilots)
-            return f"{alive} pilots, {active} active, {available} unassigned"
+            on_leave = len(squadron.pilots_on_leave)
+            return f"{active} active, {available} unassigned, {on_leave} on leave"
         return ""